@@ -5,7 +5,7 @@ version: 0.0.1
 homepage: https://github.com/wixtime/age
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
